+++ src/App.tsx
@@ -4,6 +4,7 @@ import GoodsDetailPage from '@/pages/GoodsDetailPage'
 import StoreDetailPage from '@/pages/StoreDetailPage'
 import ReceiptDetailPage from '@/pages/ReceiptDetailPage'
 import ReceiptCorrectionPage from '@/pages/ReceiptCorrectionPage'
+import PendingReceiptsPage from '@/pages/PendingReceiptsPage'
 
 export default function App() {
   return (
@@ -12,6 +13,7 @@ export default function App() {
       <Route path="/goods" element={<GoodsListPage />} />
       <Route path="/goods/:id" element={<GoodsDetailPage />} />
       <Route path="/stores/:storeId" element={<StoreDetailPage />} />
+      <Route path="/receipts/pending" element={<PendingReceiptsPage />} />
       <Route path="/receipts/:receiptId/correct" element={<ReceiptCorrectionPage />} />
       <Route path="/receipts/:receiptId" element={<ReceiptDetailPage />} />
     </Routes>